force integer value for `np.arange` when indexing h5 file with 4D and 5D data  be8830

+++ suite2p/io/h5.py
@@ -74,7 +74,7 @@ def h5py_to_binary(ops):
                         im = f[key][irange, :, :]
                     else:
                         irange = np.arange(
-                            ik / ncp, min(ik / ncp + nbatch / ncp, nframes_all / ncp),
+                            int(ik / ncp), int(min(ik / ncp + nbatch / ncp, nframes_all / ncp)),
                             1)
                         if irange.size == 0:
                             break
